refactor so stream formats are not considered a platform
