+++ homeassistant/components/stream/const.py
@@ -1,0 +1,9 @@
+"""Constants for Stream component."""
+DOMAIN = 'stream'
+
+ATTR_ENDPOINTS = 'endpoints'
+ATTR_STREAMS = 'streams'
+
+CONF_KEEPALIVE = 'keepalive'
+
+OUTPUT_FORMATS = ['hls']
